+++ crates/wasmtime/src/engine.rs
@@ -1,0 +1,148 @@
+use crate::Config;
+use std::sync::Arc;
+#[cfg(feature = "cache")]
+use wasmtime_cache::CacheConfig;
+use wasmtime_jit::Compiler;
+use wasmtime_runtime::debug_builtins;
+
+/// An `Engine` which is a global context for compilation and management of wasm
+/// modules.
+///
+/// An engine can be safely shared across threads and is a cheap cloneable
+/// handle to the actual engine. The engine itself will be deallocate once all
+/// references to it have gone away.
+///
+/// Engines store global configuration preferences such as compilation settings,
+/// enabled features, etc. You'll likely only need at most one of these for a
+/// program.
+///
+/// ## Engines and `Clone`
+///
+/// Using `clone` on an `Engine` is a cheap operation. It will not create an
+/// entirely new engine, but rather just a new reference to the existing engine.
+/// In other words it's a shallow copy, not a deep copy.
+///
+/// ## Engines and `Default`
+///
+/// You can create an engine with default configuration settings using
+/// `Engine::default()`. Be sure to consult the documentation of [`Config`] for
+/// default settings.
+#[derive(Clone)]
+pub struct Engine {
+    inner: Arc<EngineInner>,
+}
+
+struct EngineInner {
+    config: Config,
+    compiler: Compiler,
+}
+
+impl Engine {
+    /// Creates a new [`Engine`] with the specified compilation and
+    /// configuration settings.
+    pub fn new(config: &Config) -> Engine {
+        debug_builtins::ensure_exported();
+        Engine {
+            inner: Arc::new(EngineInner {
+                config: config.clone(),
+                compiler: config.build_compiler(),
+            }),
+        }
+    }
+
+    /// Returns the configuration settings that this engine is using.
+    pub fn config(&self) -> &Config {
+        &self.inner.config
+    }
+
+    pub(crate) fn compiler(&self) -> &Compiler {
+        &self.inner.compiler
+    }
+
+    #[cfg(feature = "cache")]
+    pub(crate) fn cache_config(&self) -> &CacheConfig {
+        &self.config().cache_config
+    }
+
+    /// Returns whether the engine `a` and `b` refer to the same configuration.
+    pub fn same(a: &Engine, b: &Engine) -> bool {
+        Arc::ptr_eq(&a.inner, &b.inner)
+    }
+}
+
+impl Default for Engine {
+    fn default() -> Engine {
+        Engine::new(&Config::default())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{Config, Engine, Module, OptLevel};
+    use anyhow::Result;
+    use tempfile::TempDir;
+
+    #[test]
+    fn cache_accounts_for_opt_level() -> Result<()> {
+        let td = TempDir::new()?;
+        let config_path = td.path().join("config.toml");
+        std::fs::write(
+            &config_path,
+            &format!(
+                "
+                    [cache]
+                    enabled = true
+                    directory = '{}'
+                ",
+                td.path().join("cache").display()
+            ),
+        )?;
+        let mut cfg = Config::new();
+        cfg.cranelift_opt_level(OptLevel::None)
+            .cache_config_load(&config_path)?;
+        let engine = Engine::new(&cfg);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 0);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 1);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+
+        let mut cfg = Config::new();
+        cfg.cranelift_opt_level(OptLevel::Speed)
+            .cache_config_load(&config_path)?;
+        let engine = Engine::new(&cfg);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 0);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 1);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+
+        let mut cfg = Config::new();
+        cfg.cranelift_opt_level(OptLevel::SpeedAndSize)
+            .cache_config_load(&config_path)?;
+        let engine = Engine::new(&cfg);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 0);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+        Module::new(&engine, "(module (func))")?;
+        assert_eq!(engine.config().cache_config.cache_hits(), 1);
+        assert_eq!(engine.config().cache_config.cache_misses(), 1);
+
+        // FIXME(#1523) need debuginfo on aarch64 before we run this test there
+        if !cfg!(target_arch = "aarch64") {
+            let mut cfg = Config::new();
+            cfg.debug_info(true).cache_config_load(&config_path)?;
+            let engine = Engine::new(&cfg);
+            Module::new(&engine, "(module (func))")?;
+            assert_eq!(engine.config().cache_config.cache_hits(), 0);
+            assert_eq!(engine.config().cache_config.cache_misses(), 1);
+            Module::new(&engine, "(module (func))")?;
+            assert_eq!(engine.config().cache_config.cache_hits(), 1);
+            assert_eq!(engine.config().cache_config.cache_misses(), 1);
+        }
+
+        Ok(())
+    }
+}
